auth-service/requirements.txt
Django>=4.2.0,<5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
django-cors-headers==4.3.0
python-dotenv==1.0.0